Fill area with black before drawing VR distortion effect
The window manager will draw underneath our window, and then we're expected to paint
over it -- if we just ClearRect(), all that does is make our surface transparent
so that when we blend into the window, the bogus decorations still show through.
We need to do an opaque fill with black.